pyproject.toml
[project]
name = "books-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.124.4",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "psycopg2>=2.9.11",
]
